--- client/src/components/App.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-import FrontPage from './frontpage/FrontPage'
-
-
-class App extends Component {
-  render() {
-    return (
-      <div className='container'>
-        <FrontPage />
-      </div>
-    )
-  }
-}
-
-export default App
