+++ client/src/redux/actions/loginActions.js
@@ -1,0 +1,10 @@
+import { createAction } from 'redux-actions'
+
+export const login = createAction('LOGIN', (username, password) => ({ username, password }))
+export const loginSuccess = createAction('LOGIN_SUCCESS')
+export const loginFailure = createAction('LOGIN_FAILURE')
+
+export const logout = createAction('LOGOUT')
+export const logoutSuccess = createAction('LOGOUT_SUCCESS')
+export const logoutFailure = createAction('LOGOUT_FAILURE')
+export const emptyTokenExpired = createAction('EMPTY_TOKEN_EXPIRED')
