+++ client/src/redux/reducers/loginReducer.js
@@ -3,8 +3,6 @@ import * as actions from '../actions/actions'
 
 const initialState = {
   loggedIn: null,
-  loginError: null,
-  logoutError: null,
   tokenExpired: false
 }
 
@@ -12,21 +10,12 @@ const reducer = handleActions(
   {
     [actions.loginSuccess]: (state, action) => ({
       ...state,
-      loggedIn: action.payload,
-      loginError: null
+      loggedIn: action.payload
     }),
     [actions.logoutSuccess]: (state, action) => ({
       ...state,
       loggedIn: null,
       tokenExpired: action.payload
-    }),
-    [actions.loginFailure]: (state, action) => ({
-      ...state,
-      loginError: action.payload
-    }),
-    [actions.logoutFailure]: (state, action) => ({
-      ...state,
-      logoutError: action.payload
     }),
     [actions.emptyTokenExpired]: (state) => ({
       ...state,
