+++ client/src/redux/reducers/reducers.js
@@ -1,0 +1,13 @@
+import loginReducer from './loginReducer'
+import generalReducer from './generalReducer'
+import errorReducer from './errorReducer'
+import applicationReducer from './applicationReducer'
+import jobPostingReducer from './jobPostingReducer'
+
+export const reducers = {
+  loginReducer,
+  jobPostingReducer,
+  errorReducer,
+  applicationReducer,
+  generalReducer
+}
