+++ client/src/redux/sagas/jobPostingSaga.js
@@ -1,6 +1,7 @@
 import { delay } from 'redux-saga'
 import { takeLatest, takeEvery, put, call, select } from 'redux-saga/effects'
 import * as actions from '../actions/actions'
+import * as selectors from '../selectors/selectors'
 import jobPostingApi from '../apis/jobPostingApi'
 
 function* submitJobPosting({ payload }) {
@@ -13,7 +14,7 @@ function* submitJobPosting({ payload }) {
       showFrom: payload.showFrom,
       showTo: payload.showTo
     }
-    const recruiter = yield select(getCurrentUser)
+    const recruiter = yield select(selectors.getUser)
 
     if (!recruiter) {
       return
@@ -40,7 +41,7 @@ function* submitJobPosting({ payload }) {
 
     if (response.status === 201 || response.status === 200) {
       yield put(actions.addJobPostingSuccess())
-      yield call(fetchJobPostings)
+      yield put(actions.fetchJobPostings())
       yield delay(5000)
       yield put(actions.removeJobPostingStatus())
     }
@@ -54,31 +55,9 @@ function* submitJobPosting({ payload }) {
   }
 }
 
-function* fetchJobPostings({ payload }) {
-  try {
-    const recruiter = payload.recruiter
-
-    const response = yield call(jobPostingApi.get, { recruiter })
-
-    if (response.status === 200) {
-      yield put(actions.setJobPostings(response.data))
-    }
-  } catch (e) {
-    console.log('Could not fetch job postings')
-  }
-}
-
-function* addNewStageForJobPosting({ payload }) {
-  yield put(actions.addNewStageForJobPosting(payload.newStage))
-}
-
-function* removeStageInJobPosting({ payload }) {
-  yield put(actions.removeStageInJobPosting(payload.stage))
-}
-
 function* fetchJobPosting({ payload }) {
   try {
-    const recruiter = yield select(getCurrentUser)
+    const recruiter = yield select(selectors.getUser)
     const token = recruiter ? recruiter.token : null
     const id = payload.postingId
 
@@ -98,7 +77,7 @@ function* fetchJobPosting({ payload }) {
 function* fetchJobPostingWithStages({ payload }) {
   try {
     yield delay(10)
-    const recruiter = yield select(getCurrentUser)
+    const recruiter = yield select(selectors.getUser)
 
     if (!recruiter) {
       return
@@ -138,7 +117,7 @@ function* fetchJobPostingWithStages({ payload }) {
 
 function* fetchJobPostingApplicants({ payload }) {
   try {
-    const recruiter = yield select(getCurrentUser)
+    const recruiter = yield select(selectors.getUser)
 
     if (!recruiter) {
       return
@@ -150,7 +129,7 @@ function* fetchJobPostingApplicants({ payload }) {
 
     if (response.status === 200) {
       const jobApplicants = response.data
-      yield put(actions.fetchApplicantsSuccess(jobApplicants))
+      yield put(actions.updateStages(jobApplicants))
     }
   }
   catch (e) {
@@ -160,9 +139,9 @@ function* fetchJobPostingApplicants({ payload }) {
 
 function* updateJobPostingStages({ payload }) {
   try {
-    const recruiter = yield select(getCurrentUser)
-    const stages = yield select(getCurrentStages)
-    const posting = yield select(getCurrentJobPosting)
+    const recruiter = yield select(selectors.getUser)
+    const stages = yield select(selectors.getStages)
+    const posting = yield select(selectors.getJobPosting)
 
     /*
     First remove the moved stage from old index with splice.
@@ -180,7 +159,7 @@ function* updateJobPostingStages({ payload }) {
       return { ...stage, orderNumber: index }
     })
 
-    yield put(actions.moveStageSuccess(reOrderedStages))
+    yield put(actions.updateStages(reOrderedStages))
 
     const jobPosting = { ...posting, stages: reOrderedStages }
     delete jobPosting.isHidden
@@ -202,20 +181,13 @@ function* updateJobPostingStages({ payload }) {
   }
   catch (e) {
     console.log(e)
-    yield put(actions.moveStageSuccess(payload.oldStages))
+    yield put(actions.updateStages(payload.oldStages))
   }
 }
 
-export const getCurrentUser = state => state.loginReducer.loggedIn
-export const getCurrentStages = state => state.postingReducer.stages
-export const getCurrentJobPosting = state => state.postingReducer.jobPosting
-
-export const watchFetchJobPostings = takeLatest(actions.fetchJobPostings().type, fetchJobPostings)
 export const watchFetchJobPosting = takeLatest(actions.fetchJobPosting().type, fetchJobPosting)
 export const watchFetchJobPostingWithStages = takeLatest(actions.fetchJobPostingWithStages().type, fetchJobPostingWithStages)
 export const watchSubmitJobPosting = takeEvery(actions.submitJobPosting().type, submitJobPosting)
-export const watchNewStageToJobPosting = takeEvery(actions.addNewStageForJobPosting().type, addNewStageForJobPosting)
-export const watchRemoveStageInJobPosting = takeEvery(actions.removeStageInJobPosting().type, removeStageInJobPosting)
 export const watchMoveJobPostingStage = takeEvery(actions.moveStage().type, updateJobPostingStages)
 export const watchFetchApplicants = takeEvery(
   actions.fetchApplicants().type,
