+++ client/src/redux/selectors/applicationSelectors.js
@@ -1,0 +1,2 @@
+export const getComments = state => state.applicationReducer.comments
+export const getApplicationStatus = state => state.applicationReducer.applicationSuccessful
