+++ client/src/redux/selectors/errorSelectors.js
@@ -1,0 +1,4 @@
+export const getErrorMessage = state => state.errorReducer.errorMessage
+export const getLoginError = state => state.errorReducer.loginError
+export const getLogoutError = state => state.errorReducer.logoutError
+export const getStageError = state => state.errorReducer.stageError
