+++ client/src/redux/selectors/generalSelectors.js
@@ -1,0 +1,2 @@
+export const getPostings = state => state.generalReducer.jobPostings
+export const getStages = state => state.generalReducer.stages
