+++ client/src/redux/selectors/jobPostingSelectors.js
@@ -1,0 +1,7 @@
+export const getShowTo = state => state.jobPostingReducer.showTo
+export const getShowFrom = state => state.jobPostingReducer.showFrom
+export const getCreationRequestStatus = state => state.jobPostingReducer.creationRequestStatus
+export const getDefaultStageNames = state => state.jobPostingReducer.defaultStageNames
+export const getSetTimespan = state => state.jobPostingReducer.setTimespan
+export const getJobPostingStages = state => state.jobPostingReducer.jobPostingStages
+export const getJobPosting = state => state.jobPostingReducer.jobPosting
