+++ client/src/redux/selectors/loginSelectors.js
@@ -1,0 +1,2 @@
+export const getUser = state => state.loginReducer.loggedIn
+export const getTokenExpiredStatus = state => state.loginReducer.tokenExpired
